+++ examples/sse/src/app/sse-finite/route.ts
