+++ examples/sse/src/app/sse-infinite/route.ts
